feat(panel-collapsible): create panels that can be collapsed
A new component has been added to the theme that combines two existing
bootstrap features:

 * Panel - http://getbootstrap.com/components/#panels
 * Collapse - http://getbootstrap.com/javascript/#collapse

A collapsible panel can be created by adding a class of
".panel-collapsible" to your panel. This will work for the 6 bootstrap
states (primary, success, etc.)

The user can then define a target for the collapse, this should be a div
that wraps the ".panel-body"

Since the collapse and expand icons use font awesome, it has been
included as a development dependency. It is anticipated that
font-awesome will be compiled into the theme at some point in the
future.