feat(Gruntfile.js): add a release task for upload to s3
A new task `grunt release` will build the release and upload it to S3 to
the bucket defined in the `grunt-aws.json` file. A sample JSON file has
been included.

Since the theme is a modified version of bootstrap, it has been renamed
to `theme.css` instead of `bootstrap.css`