fix(input): set height multiplier to accommodate padding
The padding on input elements now ensures that select box options text
is no longer cut off.

This appeared to be fixed in 01b3788f63150307be5d9eccb3c9e9498252ec51
but it appears that the select options in `index.html` were all above
the baseline so the issue was not highlighted. The options inside the
select have been updated to values that will be cut off if the select
box is too small.