feat(panels): change header font weight and padding
Previously the header title was using a thin font which would make
reading the title hard. The font weight has been replaced  to
400 to make reading easier.

The left and right padding were previously unbalanced. There was a
higher padding on the right than on the left. Right and left padding
values have been reviewed to 12px to visually balance the title
header.