feat(table): change padding and color variable names
Previously the <th> and <tr> table elements had a padding of 8px
on normal mode and 6px on condensed mode. To match the paddings
visually with other elements of the theme, the paddings were
changed to 8px top and bottom and 12 px left and right on normal
mode and to 6px top and bottom and 12 px left and right on
condensed mode.

Color values for the table border and hover states were changed
from absolute values to the variable names used as standard
trough out the theme.