feat(navs-tabs): change colors and paddings for the tabs component
Previously the tabs component of the navs category was inconsistent
in terms of style details when compared to the rest of the theme. To
make the component more consistent with the rest of the theme
components changes were made to border colors, background colors,
paddings and margins. Also, variables are now replacing previously
hex color codes.