feat(alerts): change colors and font weights of alerts component
Previously the alerts component had a thin h4 title that would make
it hard to read. The font-weight of the h4 has therefore been
increased. The bottom margin of the h4 title has also been reviewed to
2px to enhance visual balance between title and message.

The close button of the alert box was also reviewed in terms of
initial transparency and hover color in order to enhance visibility
and contrast.

The links and close font weight variable that was
previously 'bold' has been replaced by '700' in order to generate
consistency along the code.