feat(labels): change default label colors and labels font weight
Previously the default label was not visible due to the background
and font colors being white. This has now been reviewed in order
to be consistent with the rest of the theme and to ensure
readability of the default label component.

To enhance readability the font weight of  all the labels have
been reviewed from 'bold' to '500'.