feat(progress_bars): change background color
Previously the background color of progress bars was not consistent
with the rest of the theme. The background color for progress bars
has therefore been changed to 'gray-lightest' (#f7f7f7). For the
color value, instead of a hex value, a variable name was used in
order to generate consistency along the code.