feat(jumbotron): change colors and paragraph font weight
Previously the background color of the jumbotron component was not
consistent with the rest of the theme. Color changes have been made
to the background, heading and paragraph color in order to improve
color consistency with the remaining components of the theme.

In order to improve readability of the paragraph text the font
weight has also been increased.