feat(headings): add strong and italic styles
Previously the headings had no strong and italic options
for developers to use. In order to enable a wider selection
of heading styles, a strong and italic style was included for
all the headings (h1, h2, h3, h4, h5 and h6). The styles can
be used by nesting the strong and em tags inside the heading
elements. Example:

* `<h1><strong>Heading 1</strong></h1>`
* `<h3><em>Heading 3</em></h3>`

The global fonts included in the theme were also reviewed in
order to include the italic font style that was previously
not included.