feat(help-block): reduce prominence
In order to achieve visual balance between form elements, the style
of the help-block has been changed in terms of `font-size`,
`margin-top` and `line-height`.

As an example to illustrate how the help block class would look
when applied to an element it has been added to the theme index
file.