feat(grunt): run less task as default and watch `build.less`
Previously, if a new component was added into `build.less`, it was not
reflected in the theme as it was not being watched by Grunt. So a
developer needed to restart grunt which is cumbersome, therefore,
`build.less` is now watched by Grunt for any changes.

However, this would still not update `theme.css` as the theme was not
built until Grunt noticed a change in a watched file. So a developer
was forced to make a change that wasn't required in order for the
changes to be reflected in the theme, therefore, the `less` task has
been included in the default grunt task.