fix(list-group): examples on low resolution
Previously the list-group examples didn't render well on low resolution
devices. To improve this experience the list-group titles have been
moved from a separate row to be within the same row as each list-group
example.

Additionally the media-list-group example has been wrapped in a
ls-component div, so that it renders the same as other list-group
examples.