style(list-group): variable header height on media-list-group
The `media-list-group` component would have too much space between the
header and the first paragraph if the header only contained one line.

This was because the height of the header was being set with
`min-height` which was adapted to a two line header.

Removing the min height and adding a bottom margin creates the same
space separation between the header and first paragraph elements and
looks more consistent when a single line of text is being displayed.