fix(buttons): remove `box-shadow` from active state
There is an issue on iOS 7 and iOS 8 where the active state on the `btn`
class prevents a touch event on any element that is outside the initial
height area of the iframe and causes the iframe content to scroll to the
top.

This behaviour is exhibited in the `app store` app as when trying to add
an app using the `+ Add` button (which has the .btn class) doesn't add
the app that is expected as the content scrolls to the top on touch and
then fires the touch event on the app at the top of the page instead.

A bug report has been raised with Apple under code `19043750`