fix(components): prevent edit form controls wrapping on mobile
The `inline-form` bootstrap class only provides `display: inline-block`
on a resolution greater than `767px` so the button items display
as block elements on any resolution lower than that.

The solution is to just force `display: inline-block` on the input
element so it is always set to show the input as an inline-block
element regardless of the resolution width.