refactor(components): use standard components in media-list-group
The media list group now uses more standard html and bootstrap
components to create a media list item, which makes it more
customisable when wanting to use other components within the media list
group item.

The `list-group-media-item` can also be given an `active` class now to
identify it as being the selected element.