refactor(components): remove unnecessary markup from media-badge
The `media-badge` component was reliant on adding additional markup and
classes from bootstrap to display correctly. The markup required to
created stacked icons has now been removed to make the markup cleaner
and less reliant on assumptions the user would add a first stacked icon
to be hidden.

The font badges now just take a font inside the `media-badge` component
which has a `background-color` applied to the containing div rather than
expecting to use stacked icons to provide the relevant colors.

This makes it easier to interchange media badge types with minimal
amendments to the existing markup.