fix(components): remove unnecessary margin from media-item img
The `media-item` image does not require a `margin-right` as the
`media-content` component has a fixed `margin-left`. The removal of
this style also enables the `media-badge` image component to be used
with an `img-circle` class to ensure it is rounded.

The `img-circle` class is now required on the `media-badge` for image
media badge types as this will ensure the media badge is always
circular.