fix(headings): line-height increased to be compatible with ellipsize
Previously the default line-height of headings was 1.1em, this value was
inherited directly from bootstrap.

A problem was identified with headings that had the `ellipsize` class
applied. The descenders (letters that hang low) present in such headings
would be cut off. This was a result of `overflow:hidden` being applied
via the `ellipsize` class.

The fix for this was to increase the default `line-height` of headings.