feat(list-group): error highlighting for inline forms
A new List group type has been added specifically for listing inline
forms, the css class for this type is `form-list-group`.

When an error needs to be displayed for a inline form (within a
`form-list-group`) the bootstrap `has-error` css class should be
applied to the `form-inline` element, and a `help-block` element be
included within the `form-inline` to display the error text.

When an error is displayed the form in question is displayed in-front
of any other inline forms. This is achieved by making the error
form a positioned element (relative), which by default will display it
in-front of non-positioned elements.