feat(components): add unselectable class to list group items
Previously, list group item elements had a hover state which changed
their background colour. In some occasions a list group is required that
isn't selectable but being used for styling purposes.

A new class of `unselectable` can now be applied to a `list-group-item`
or a `media-item` element which disables the hover/focus state of the
element.