feat(components): create new version of media-list-group
The Media-List-Group component has been rebuilt to make it more
versatile as it is now more composable of existing theme components,
which means it can cover a wider variety of use cases.

The left image is no longer a fixed size and can be replaced with any
size media badge (image or icon) and can contain the media badge inner
icon. The image can also be replaced with a button group to provide
action buttons per media list item.

The image or buttons can also be floated to the right of the item
content by applying the `pull-right` class instead of `pull-left`.

The content within the `media-content` element can have the line spacing
controlled by applying an additional class that controls the margin.

 * `line-tighter` (applies `margin-bottom: 0px;`)
 * `line-tight` (applies `margin-bottom: 2px;`)
 * `line-normal` (applies `margin-bottom: 5px;`)
 * `line-loose` (applies `margin-bottom: 7px;`)
 * `line-looser` (applies `margin-bottom: 10px;`)

The theme example page `index.html` has been updated to include a number
of use cases for the new Media-List-Group component.