chore(Gruntfile.js): prevent overwriting existing file
To ensure that a version of the theme is not overwritten, a new
`checkVersion` task has been created. This uses the same credentials as
the s3 task to perform a get on the object. If it succeeds then the
object exists and therefor the pipeline should not continue. If it
errors then the object does not exists.

To ensure that the `s3:development` task is not accidentally executed -
it is renamed to append the date so that it cannot be run easily from
the command line.