fix(Gruntfile.js): ensure config for s3 is available
After renaming the s3 task, the config can no longer be found as the key
is different.

The Grunt documentation[1] specifies that this should be renamed
appropriately, however this does not seem to be the case. To ensure this
works, the key is copied with the appended timestamp at the point of
initializing the config.

[1] http://gruntjs.com/api/grunt.task#grunt.task.renametask