style(components): set media-list with buttons example to unselectable
The media-list with buttons example should not provide a hover state on
the entire list item because good UI/UX practice dictates that the
buttons should be the only actionable elements inside the list item and
the list item itself should not provide any action on hover or click.

The hover state of the buttons is still available.