docs(public): update README and provide CONTRIBUTING & LICENSE
`README.md` has been updated with a background of the theme and updated
instructions. Whilst a new `CONTRIBUTING.md` has been created to provide
developers with guidelines on how to contribute to Livestax projects.
The file has been purposely been written generalized so that it can be
incorporated into other projects.

The `LICENSE` has also been added to provide developers of the
details of the code's license.