feat(site-builder): add grunt task to auto build public content
A new grunt task has been added that will auto build the new multi page
theme public website based off of html templates and html source files.

The html templates and the source files reside in the new `docs/`
directory along with a `build.json` file which acts as the build
manifest to determine which pages get built and in which order.

The main heading of the mainfest determines the html file, the following
heading determines the main title section and the subsequent headings
determine headings for the elements.

i.e.

```
  "styles": {
    "typography": [
      "headings"
    ]
  }

  styles (will generate styles.html)
  and generate a H1 title of "typography"
  with a generated H2 section for "headings"
```

The build manifest file directly relates to a directory structure within
the `docs/` directory. So `styles:typography:headings` is mapped to the
following file:

`docs/styles/typography/headings.html`

The source html files are an enhanced markup allowing for easier styling
of elements without needing to know the relevant classes. Any html is
allowed with the addition of these two tags:

  * `<title>My Title</title>`
    * Will generate a H3 tag appropriately styled.
  * `<example>my example code</example>`
    * Will generate an example container with a 'view source' button.

the `public/` directory will not be auto generated so it cannot be
deleted and is required to be checked into the repository along with any
files that are contained within it.