feat(list-group): add media-group-item
Previously the Livestax theme provided a `media-list-group` class which
was intended to be used in combination with a `list-group` class.
`list-group-item` elements within a `media-list-group` had numerous
styles applied that provided rich layout options.

One of the limitations of this was found to be that it was not possible
to use these "media items" outside of a list. Something developers may
find useful when they are composing their own components.

The original `media-list-group` class has now been removed. A new
`media-group-item` has been added which can be used in combination with
a `list-group-item` or on its own outside of a `list-group`.

When a `media-group-item` is use in combination with a `list-group-item`
the element behaves like an item in a list (selectable, hoverable etc).
This can be done inside or outside of a `list-group`.

When a `media-group-item` is used outside of a `list-group` it just
dictates media layout. The `progress-indicator` example has been updated
to demonstrate this type of use case.