refactor(form-list-group): replace with use of inline-form-item
Previously the Livestax theme provided a `form-list-group` which was
intended to be used in combination with a `list-group`. Such lists
provided the ability to create lists that incorporated "inline forms".

`form-list-group` lists didn't use `list-group-items` and as such were
unconventional. Also the use of these "inline forms" was restricted to
use within a list.

The original `form-list-group` class has now been removed. A new
`inline-form-item` has been added which can be used in combination with
a `list-group-item` or on its own outside of a `list-group`.

When an `inline-form-item` is used in combination with a
`list-group-item` it has padding and borders consistent with a
list item, with the input field being borderless. This can be done
inside or outside of a `list-group`.

When a `inline-form-item` is used independently of a `list-group-item`
then it has no padding and is borderless. The input field has
borders.