refactor(dropdown-list-group): replace with use of dropdown-item
Previously the Livestax theme provided a `dropdown-list-group` which was
intended to be used in combination with a `list-group`. Such lists
contained items with a dropdown menu.

`dropdown-list-group` lists contained `list-group-items` that had to be
wrapped within a `btn-group` and as such were unconventional.
Also the use of these "dropdown items" was restricted to use within a
list.

The original `dropdown-list-group` class has now been removed. A new
`dropdown-item` has been added which can be used in combination with
a `list-group-item` or independently.

When a `dropdown-item` is used in combination with a `list-group-item`
it behaves like a list item (borders, hoverable etc).

When a `dropdown-item` is used independently of a `list-group-item` then
it is borderless and the `anchor` tag appears as a standard link.
In such a use case the `dropdown-item` should be used in combination
with the `clearfix` css class.