feat(media-group-item): remove border top if only child
The `media-group-item` applies a top border to the first child in the
list. All items have a border bottom, except for the last child.

It was observed that lists with only one item looked slightly strange,
as there would be just a border top on the one and only item.

Lists containing only one `media-group-item` are now borderless.