refactor(media-badge): support any text content
Previously the `media-badge` component was very tied to displaying icons
or images. A requirement was identified to display any type of text
content, for example to display a percentage number.

The `media-badge` component has been refactored to expect text content
to be wrapped in a text tag (h1-h6, p), and it is this that dictates the
size of the content.

The `media-badge` now utilises `flex-box` to accomplish vertical
alignment of content.

Documentation for all the different media-badge implementations
has been provided.