feat(buildDocs): auto generate side menu for doc pages
A side menu of component links is now auto generated for each page of
the theme documentation allowing for easier access to sections and a
quick overview of the page contents.

The menu becomes sticky when the page is scrolled so it is always in
view and available to use.