fix(media-group-item): remove space generated by clearfix
When a `media-group-item` contains a small amount of content, it is
possible that the badge element will be "cut off", as a result of it
being a floated element and hence taken out of the natural flow.

The recommended practice to fix this layout, is to apply the bootstrap
`clearfix` class to the `media-group-item`. However it was discovered
that applying `clearfix` produces a small amount (7px) of additional
vertical space within the `media-group-item`. To fix this issue a
negative margin bottom is applied to the `after` pseudo element of the
`clearfix`.