feat(deploy): move dependencies into public and deploy
Previously, the built doc files targeted dependencies outside the
`public` directory. This would break if servings the docs in isolation,
therefore, the build manifest (`build.json`) has been updated with a
`vendor` object. The `vendor` object is ignored in the build task, so a
new page is not created, however, the file is copied into a directory
in the `public` directory (`public/vendor/{extension}/{filename}`).

Moving the dependencies meant that the files that used them needed
updating with the correct URLs.

The task that enabled deployment to gh-pages has been removed in favour
of `grunt deploy-docs`. This task pushes the files in the public
directory to the S3 bucket specified in `grunt-aws.json`. The S3
credentials example has been updated to enable details of the
documentation to be recorded.