refactor(media-group-item): vertical align media-content
The `media-content` element of the `media-group-item` is now
`shrink-wrapped` and centered vertically.

The `media-badge` element is pinned to the top of the `media-group-item`
regardless of the amount of content within the `media-content` element.

These layout changes are achieved using flex-box. This reliance on
flex-box eliminates the use of floats for element positioning. As a
result the fix introduced in commit
8ce870dda1c6ec6878a907a179cc8e9fef7085fd for removing white-space
generated by the use of `clearfix` has been removed, as `clearfix` is no
longer required.

The `media-group-item` examples have been updated to demonstrate how
the component behaves with minimal content as well as a large amount of
content.