feat(docs): add contributing and license details
Provide detailed instructions on the public docs on how to install and
contribute to the theme. It gives a brief explanation to how the
documentation is generated and how to create new components for the
docs.

For completeness, the license information for both the docs and the
underlying code are now clearly displayed on the public docs.