feat(media-group-item): support media-content column layout
The `media-group-item` component now supports a new
`media-content-columns` class which can be used in association with the
`media-content` class.

The `media-content-columns` class prescribes that direct decedents will
be rendered in columns.  A `stretch` class can be applied to any of
these decedent, with the effect being to expand the width of the
element until it uses all "unused" space within the parent element.
Typically you would expect Developers to only apply `stretch` to one of
the column elements.

The `media-group-item` examples have been updated to include
`media-content-columns` cases.