refactor(media-group-item): control layout with `.media-row`
Previously the positioning of the children of a `media-group-item` was
controlled through a combination of css floats and flex-box.
In preparation for collapsible list items this layout mechanism has been
refactored.

A new `media-row` css class is available to be used in association with
a `media-group-item`, or applied to a nested container within a
`media-group-item`.

Direct descendants of a `media-row` are rendered side-by-side in a row.
All content within a `media-group-item` (including `media-badge`) now
sits within a `media-row`.

A `stretch` class can be applied to children of a `media-row`, with
the effect of expanding the width of the child until it uses all
"unused" space within the parent element. Typically you would expect
Developers to only apply `stretch` to one of the children.

The `media-content` and `media-content-columns` classes have been
removed as they are no longer required to control layout.