fix(list-groups): don't set first-item border if item is active
Previously the `list-groups` bootstrap overrides always set the first
item of the list to have a gray border.

This caused a noticeable render glitch when the first-item was active
and in a hover state. Namely that the border-top color changed to
blue giving the illusion that the first-item increased in size.

This solution ensures that the first-item doesn't have a border-top if
it is active.