feat(list-groups): add condensed list-group-item class
The `.condensed` class reduces the padding on a list group item to make
it work better in conjunction with a header list group item (using the
`.list-group-item-muted` class applied).

The `.list-group-item-muted` class is also now applied to all `li`
elements and the text colour is now a darker grey.

The occasional `border-top` on the `list-group-item` has also been
reworked so a border always appears on the first visible item in the
list in case some `li` elements are hidden.