feat(docs): add ability to provide raw example code
When creating examples that may use inline JavaScript, the DOM may be
altered meaning the example code no longer resembles the original code
when it is displayed in the Source Modal.

It is now possible to include the same code twice, one for parsing by
the page and one for displaying in the Source Modal.

Any html surrounded by the `<pagehtml></pagehtml>` tags will be
available for inline JavaScript to process and it will be displayed on
the documentation page but it will be removed from the Source Modal so
any processed html will not show as example source code.

Any html surrounded by the `<docshtml></docshtml>` tags will not be
displayed on the documentation page but will be shown in the Source
Modal as the example source code.