feat(examples): add search lists example to examples page
A new Examples page has been added that will display examples of
component composition and usage.

The first example is a Search List which utilises a Search Input and
List Group component.

JavaScript is provided in-line to demonstrate the functionality of the
Search List and the script also shows up in the example Source Modal for
ease of access.