chore(components): update basic list group example to use ul/li
Previously, the basic list group example used a `div` and a collection
of anchor tags to demonstrate a list group. This was not consistent with
other list examples that use a `ul` and a collection of `li` elements to
demonstrate the list.

In utilising a `ul` and `li` for the example, it exposed another issue
that list elements were not highlighted by default on hover so this has
been fixed and to prevent the hover state the `.unselectable` class is
applied as it is supposed to be in the other list group elements.

This additional fix caused some conflicts in other list components so
they were fixed accordingly.