fix(components): update progress indicator to use new media-group item
The media-group item has changed in markup so elements are no longer
floated, this caused an issue with the progress indicator as it used a
media-group item to lay it out with the badge.

The new media-group markup has now been used on the progress indicator
examples to display it correctly.