feat(collapsible-item): add no-indicator option
If the `no-indicator` css class is applied to a `collapsible-item`
component then the collapsible indicator icon for that item will not be
visible.

Note that the `no-indicator` rule does not cascade to descendent
`collapsible-item`s. That is the `no-indicator` must be explicitly
added to any child `collapsible-item`s if this behaviour is desired.