feat(components): add border classes to list-groups
By default, the list groups will no longer have a side border or an
absolute top or bottom border. This ensures that all list group types
look aesthetically correct within apps.

However, sometimes it is required to have a list border top or bottom,
and in certain cases it may be required to have a left/right border as
well. To accommodate this requirement, the following classes have been
added that need to be applied alongside the `.list-group` class:

  * `.list-group-border-top` -> applies a top border to the list
  * `.list-group-border-bottom` -> applies a bottom border to the list
  * `.list-group-border-vertical` -> applies a side border to the list
  * `.list-group-border-all` -> applies all the above styles