chore(list-group): remove hard font-color dependency from muted style
The `list-group-item-muted` no longer has a hard coded font colour and
just changes the background of the list group item to a muted colour.

This enables more flexibility in the component as different colours can
be applied if required. If the previous font colour is required then a
`text-muted-dark` class can be applied to the list group item to provide
the same style.

The examples have been updated to show the new changes.