feat(list-group-with-indicators): apply to lists with collapsible-item's
Previously collapsible items that contained an indicator received
additional padding to ensure that the content in these elements didn't
overlap the indicator icon. This solution worked will for items with an
indicator, however it was realised that lists can contain both
collapsible and non-collapsible items, and that it's important for the
padding-right to be consistent across these items.

collapsible items no longer receive additional padding, instead a new
`list-group-with-indicators` class should be applied to any `list-group`
that contains collapsible items. This class ensures that all list types
are appropriately padded to avoid overlapping the indicator icon.