fix(collapsible-item): reset non active item text color
Commit 7ec5860f set the color of non-active `collapsible-item` elements
to be `initial`, which had the effect of setting the `color` property
to its default value, which in the case of most browsers is black
(#000000).

From a Bootstrap perspective this behaviour was not desirable due to the
fact that the `body` element is styled to explicitly set the `color`
to `@text-color` less variable (which is not black).

The ideal solution would be to set the color of non-active items to be
`inherit` to make sure the element takes its cues from the nearest
parent. However this is problematic if the nearest parent has an
`active` class applied. As in these situations we don't want the active
text color to cascade.

Therefore in response to the fact that neither `initial` or `inherit`
are suitable candidates this commit takes the following approach:

 * Nested `list-group-item` elements that are not `active` have their
   color set to the default theme text color, unless the item in
   question has a text status color explicitly applied to it.

To accomplish this task the text-status classes have extracted into a
reusable less mixin.