fix(list-groups): update unselectable class to have solid background
The `.unselectable` class requires a solid background color rather than
using `transparent` otherwise in nested lists, the child list item will
inherit the parent background color on hover.

As the Theme is opinionated about the color of the canvas being white,
then it is acceptable to enforce the hover state to be white.