fix(media-row): add inline widths for items that have wrapping text
Because of the way partially compatible browsers implement flexbox means
that in IE10, IE11 and Safari the text wrapping is disabled on elements
when they are aligned with flexbox. The action of these incompatible
browsers is to just take up as much horizontal width required to display
the text, which causes the element to take up too much room within the
media-row.

The solution is to provide an inline width on the element with the text
that requires wrapping to enforce it to have a defined width to allow
for text wrapping.